BOT_1/requirements.txt
python-telegram-bot[job-queue]==21.1.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
cryptography==42.0.5
python-dotenv==1.0.0
redis==5.0.3
apscheduler==3.10.4
pydantic==2.7.1          # Добавлена обязательная базовая библиотека
pydantic-settings==2.2.1 # Явное указание недостающего пакета